# Code/gamesUP/docs/openapi.yaml
openapi: 3.0.1
info:
  title: GamesUP API
  description: |
    API REST du catalogue et des parcours clients GamesUP.
    Les ressources personnelles sont toujours déterminées par le jeton JWT :
    aucune route ne permet de sélectionner un autre utilisateur.
  contact:
    name: GamesUP
  license:
    name: Projet pédagogique
  version: 1.0.0
servers:
- url: /
  description: Serveur courant
...
paths:
...
  /api/v1/orders:
    get:
      tags:
      - Orders
      summary: Lister ses commandes
      operationId: findCurrentUserOrders
      parameters:
      - name: page
        in: query
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          maximum: 100
          minimum: 1
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseOrderResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Règle métier empêchant l'opération.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    post:
      tags:
      - Orders
      summary: Créer une commande
      description: Le corps accepte uniquement les identifiants de jeux et quantités
        ; tout prix reçu est ignoré.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Règle métier empêchant l'opération.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/admin/users/{userId}/role:
    patch:
      tags:
      - User administration
      summary: Modifier un rôle
      description: Un administrateur ne peut pas retirer son propre rôle ADMIN.
      operationId: changeRole
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserRoleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  /api/v1/admin/users/{userId}/enabled:
    patch:
      tags:
      - User administration
      summary: Activer ou désactiver un compte
      description: Un administrateur ne peut pas désactiver son propre compte.
      operationId: changeEnabled
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserEnabledRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  /api/v1/admin/reviews/{reviewId}/visibility:
    patch:
      tags:
      - Review administration
      summary: Masquer ou republier un avis
      operationId: changeVisibility
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminReviewVisibilityRequest"
        required: true
      responses:
        "204":
          description: Visibilité modifiée.
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/orders/{orderId}:
    get:
      tags:
      - Orders
      summary: Consulter une de ses commandes
      operationId: findCurrentUserOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Règle métier empêchant l'opération.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/games/{id}:
    get:
      tags:
      - Games
      summary: Consulter un jeu actif
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GameDetailResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/admin/users:
    get:
      tags:
      - User administration
      summary: Lister les utilisateurs
      operationId: findAll
      parameters:
      - name: page
        in: query
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          maximum: 100
          minimum: 1
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseUserResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  /api/v1/admin/users/{userId}:
    get:
      tags:
      - User administration
      summary: Consulter un utilisateur
      operationId: findById_1
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    OrderCreateRequest:
      required:
      - lines
      type: object
      properties:
        lines:
          maxItems: 100
          minItems: 0
          type: array
          items:
            $ref: "#/components/schemas/OrderLineRequest"
    OrderLineRequest:
      required:
      - gameId
      - quantity
      type: object
      properties:
        gameId:
          type: integer
          description: Identifiant d'un jeu actif.
          format: int64
          example: 12
        quantity:
          minimum: 1
          type: integer
          format: int32
          example: 2
    OrderLineResponse:
      type: object
      properties:
        gameId:
          type: integer
          format: int64
        gameName:
          type: string
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
        lineTotal:
          type: number
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - PENDING
          - PAID
          - SHIPPED
          - DELIVERED
          - CANCELLED
          - ARCHIVED
        total:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineResponse"
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum:
          - CLIENT
          - ADMIN
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminUserRoleRequest:
      required:
      - role
      type: object
      properties:
        role:
          type: string
          enum:
          - CLIENT
          - ADMIN
    AdminUserEnabledRequest:
      required:
      - enabled
      type: object
      properties:
        enabled:
          type: boolean
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
    AdminReviewVisibilityRequest:
      required:
      - hidden
      type: object
      properties:
        hidden:
          type: boolean
...
    PageResponseOrderResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    GameDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
        minPlayers:
          type: integer
          format: int32
        maxPlayers:
          type: integer
          format: int32
        minAge:
          type: integer
          format: int32
        durationMinutes:
          type: integer
          format: int32
        editionNumber:
          type: integer
          format: int32
        publisher:
          type: string
        authors:
          type: array
          items:
            type: string
        categories:
          type: array
          items:
            type: string
...
    PageResponseUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
  securitySchemes:
    bearerAuth:
      type: http
      description: Jeton retourné par l'inscription ou la connexion.
      scheme: bearer
      bearerFormat: JWT